Conflicts:
	Sequential/src/Simulator.cpp

+++ Sequential/src/main.cpp
@@ -26,12 +26,12 @@ int main(int argc, const char *argv[])
     // extract the arguments
     int agents = arguments.get<int>("agents");
     int steps  = arguments.get<int>("steps");
-    double wc =  arguments.get<double>("wc");
-    double wa =  arguments.get<double>("wa");
-    double ws =  arguments.get<double>("ws");
-    double rc =  arguments.get<double>("rc");
-    double ra =  arguments.get<double>("ra");
-    double rs =  arguments.get<double>("rs");
+    float wc =  arguments.get<float>("wc");
+    float wa =  arguments.get<float>("wa");
+    float ws =  arguments.get<float>("ws");
+    float rc =  arguments.get<float>("rc");
+    float ra =  arguments.get<float>("ra");
+    float rs =  arguments.get<float>("rs");
 
     Simulator simulator(agents,steps,wc,wa,ws,rc,ra,rs);
     simulator.run();
